Refactor nixenv parser to mapxml
