+++ ee/tables/execparsers/mapxml/parse_test.go
@@ -1,7 +1,7 @@
-//go:build darwin || linux
-// +build darwin linux
+//go:build darwin
+// +build darwin
 
-package nixenv
+package mapxml
 
 import (
 	"bytes"
