+++ ee/tables/execparsers/mapxml/test-data/single-install.txt
